added confirmation alert for deleting posts

+++ admin/includes/viewAllPosts.php
@@ -109,7 +109,7 @@ if (isset($_POST['checkBoxArray'])){
             echo "<td>{$postDate}</td>";
             echo "<td><a href='../post.php?p_id={$postId}'>View Post</td>";
             echo "<td><a href='posts.php?source=edit_post&p_id={$postId}'>Edit</td>";
-            echo "<td><a href='posts.php?delete={$postId}'>Delete</td>";
+            echo "<td><a onClick=\"javascript: return confirm('Are you sure you want to delete this post?'); \"href='posts.php?delete={$postId}'>Delete</td>";
             echo "</tr>";
         }
         ?>
